Add new features to the logic analyser
Use ctrl-right and ctrl-left to go to the next transition on a user-
selected channel.

Display the user-selected channel and the magnification level (zoom).

Show a minimap - a squashed verion of the all logic channels, without
measurements, zoomed to fit across the screen, and with a indicator to
show which section of the main channel plots are shown.