Fix issue with number of pins to capture
Reverse the logic to generate the buffer size from the number of
samples and number of pins to sample in order to calculate the number
of samples in the same buffer with but a different number of pins. Had
to miss out one of the steps in the logic, for some reason, bit it
seems to work, although probably ought to find out why.