Fix bug in vsync2.pio
The bug resulted in 34 lines during the vertical backporch area, instead
of 33, and as a result 9 lines during the vertical frontporch, instead
of 10.
Added a vertical backporch measurement to the `measure` function.
Added a new trigger type (F.POR) to see the vertical frontporch more
easily.
Tweaked toolbar UI appearance.