Add CSYNC
CSYNC (Composite Sync) is a combination of HSYNC and VSYNC and is
connected to the monitor's HSYNC pin. Leave HSYNC and VSYNC for now
we can use trigger types VSYNC, RGB, and VFPOR when capturing VGA.

Also add a readme.md, a couple of shell scripts to help with flashing
the target device using openocd, and a pico_sdk_import.cmake file.