Improve trace colours
Each GPIO can have a fixed colour. Currently they repeat every 10 GPIO.
ie GP0 is the same colour as GP10, GP20 & GP30 and, for a RP235xB, GP40.
They follow the colours of the jumper jerky that I've been using,
although black was a little tricky to see, so it's a dark grey.